HID: wacom: generic: Avoid reporting a serial of '0' to userspace
commit ab41a31dd5e2681803642b6d08590b61867840ec upstream.

The xf86-input-wacom driver does not treat '0' as a valid serial
number and will drop any input report which contains an
MSC_SERIAL = 0 event. The kernel driver already takes care to
avoid sending any MSC_SERIAL event if the value of serial[0] == 0
(which is the case for devices that don't actually report a
serial number), but this is not quite sufficient.
Only the lower 32 bits of the serial get reported to userspace,
so if this portion of the serial is zero then there can still
be problems.

This commit allows the driver to report either the lower 32 bits
if they are non-zero or the upper 32 bits otherwise.

Signed-off-by: Jason Gerecke <jason.gerecke@wacom.com>
Signed-off-by: Tatsunosuke Tobita <tatsunosuke.tobita@wacom.com>
Fixes: f85c9dc678a5 ("HID: wacom: generic: Support tool ID and additional tool types")
CC: stable@vger.kernel.org # v4.10
Signed-off-by: Jiri Kosina <jkosina@suse.com>
Signed-off-by: Greg Kroah-Hartman <gregkh@linuxfoundation.org>
(cherry picked from commit 2303e0b4005c13c0faa6562e32886ebf8ee04f57)
[vegard: fix trivial conflict in context due to missing commit
 4affc2331a70fff3d0d0e8f28ead80aa2b8b589a ("HID: wacom: generic: Leave
 tool in prox until it completely leaves sense")]
Signed-off-by: Vegard Nossum <vegard.nossum@oracle.com>
(cherry picked from commit ace69fc9316dfc4a4923ea08a3d64081eaf06a02)